all test durations measured
